+++ example/lib/main.dart
@@ -6,6 +6,9 @@ import 'package:flutter/material.dart';
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
+  final GlobalKey<FetchMoreBuilderState> _fetchMoreController =
+      GlobalKey<FetchMoreBuilderState>();
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -18,9 +21,30 @@ class MyApp extends StatelessWidget {
           title: Text('Fetch More Builder'),
         ),
         body: FetchMoreBuilder(
+          fetchMoreController: _fetchMoreController,
           itemBuilder: _itemBuilder,
           dataFetcher: _dataFetcher,
           limit: 10,
+        ),
+        floatingActionButton: Row(
+          crossAxisAlignment: CrossAxisAlignment.end,
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            FloatingActionButton(
+              child: Icon(Icons.refresh),
+              onPressed: () {
+                _fetchMoreController.currentState.refresh();
+              },
+            ),
+            SizedBox(width: 15,),
+            FloatingActionButton(
+              child: Icon(Icons.get_app),
+              onPressed: () {
+                _fetchMoreController.currentState.fetch();
+              },
+            ),
+          ],
         ),
       ),
     );