+++ lib/default_bottom_loader.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class DefaultBottomLoader extends StatelessWidget {
+  const DefaultBottomLoader();
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      alignment: Alignment.center,
+      child: Center(
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: SizedBox(
+            width: 33,
+            height: 33,
+            child: CircularProgressIndicator(
+              strokeWidth: 1.5,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}