+++ lib/default_refresh_loader.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class DefaultRefreshLoader extends StatelessWidget {
+  const DefaultRefreshLoader();
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: CircularProgressIndicator(),
+    );
+  }
+}